更新页面交互

+++ app/src/main/java/com/example/baidumapdemo/wangjiaxin/Bean/museumInfo.java
@@ -19,9 +19,9 @@ public class museumInfo {
     private double lat;
     private String picture;
     private String city;
-    private int exhibition_score;
-    private int environment_score;
-    private int service_score;
+    private double exhibition_score;
+    private double environment_score;
+    private double service_score;
     private double general_score;
     private int annual_exhibitions;
     private String ing;
@@ -74,24 +74,24 @@ public class museumInfo {
         return city;
     }
 
-    public void setExhibition_score(int exhibition_score) {
+    public void setExhibition_score(double exhibition_score) {
         this.exhibition_score = exhibition_score;
     }
-    public int getExhibition_score() {
+    public double getExhibition_score() {
         return exhibition_score;
     }
 
-    public void setEnvironment_score(int environment_score) {
+    public void setEnvironment_score(double environment_score) {
         this.environment_score = environment_score;
     }
-    public int getEnvironment_score() {
+    public double getEnvironment_score() {
         return environment_score;
     }
 
-    public void setService_score(int service_score) {
+    public void setService_score(double service_score) {
         this.service_score = service_score;
     }
-    public int getService_score() {
+    public double getService_score() {
         return service_score;
     }
